+++ frontend/src/app/components/Footer.tsx
@@ -2,6 +2,12 @@
 
 import { useTheme } from "../context/ThemeContext";
 
+/**
+ * Footer Component
+ *
+ * Renders the application footer with copyright information.
+ * Adapts styling based on the current theme.
+ */
 export const Footer = () => {
   const { theme } = useTheme();
 
@@ -10,10 +16,7 @@ export const Footer = () => {
       style={{
         backgroundColor: theme === "dark" ? "#121212" : "#f9fafb",
         color: theme === "dark" ? "#9ca3af" : "#6b7280",
-        borderTop:
-          theme === "dark"
-            ? "1px solid rgba(75, 85, 99, 0.3)"
-            : "1px solid #f3f4f6",
+        borderTop: theme === "dark" ? "none" : "1px solid #f3f4f6",
       }}
       className="py-8"
     >
